deployment/requirements.txt
Flask==2.0.2
numpy==1.21.2
sklearn==0.0
PyYAML==6.0